Moved filter recently used items method to utility service
